+++ vrp-pragmatic/src/format/problem/problem_reader.rs
@@ -147,12 +147,11 @@ fn read_reserved_times_index(api_problem: &ApiProblem, fleet: &CoreFleet) -> Res
                 .iter()
                 .flat_map(|data| data.iter())
                 .map(|(_, _, time, duration)| match time {
-                    VehicleRequiredBreakTime::ExactTime(time) => {
-                        let time = parse_time(time);
-                        TimeSpan::Window(TimeWindow::new(time, time + duration))
+                    VehicleRequiredBreakTime::ExactTime { earliest, latest } => {
+                        TimeSpan::Window(TimeWindow::new(parse_time(earliest), parse_time(latest) + duration))
                     }
-                    VehicleRequiredBreakTime::OffsetTime(offset) => {
-                        TimeSpan::Offset(TimeOffset::new(*offset, *offset + duration))
+                    VehicleRequiredBreakTime::OffsetTime { earliest, latest } => {
+                        TimeSpan::Offset(TimeOffset::new(*earliest, *latest + duration))
                     }
                 })
                 .collect::<Vec<_>>();