+++ CHANGELOG.md
@@ -13,7 +13,9 @@ All notable changes to this project will be documented in this file.
 * improve managing packages with cargo's workspace
 * update dependencies
 * `breaking`: refactor internal models to simplify cost/telemetry handling
+* `breaking` change required break definition to support time range
 * change duration in csv import command from minutes to seconds
+
 
 ### Added